Startable class from config (#7075)
