+++ pinot-controller/src/test/java/org/apache/pinot/controller/helix/ControllerPeriodicTaskStarterStatelessTest.java
@@ -19,7 +19,6 @@
 package org.apache.pinot.controller.helix;
 
 import java.util.List;
-import org.apache.pinot.controller.ControllerConf;
 import org.apache.pinot.controller.ControllerStarter;
 import org.apache.pinot.controller.helix.core.PinotHelixResourceManager;
 import org.apache.pinot.core.periodictask.PeriodicTask;
@@ -40,7 +39,8 @@ public class ControllerPeriodicTaskStarterStatelessTest extends ControllerTest {
    * Test that controller starts up and helixResourceManager is non null before initiating periodic tasks
    */
   @Test
-  public void testHelixResourceManagerDuringControllerStart() {
+  public void testHelixResourceManagerDuringControllerStart()
+      throws Exception {
     startController();
     stopController();
   }
@@ -51,15 +51,15 @@ public class ControllerPeriodicTaskStarterStatelessTest extends ControllerTest {
   }
 
   @Override
-  protected ControllerStarter getControllerStarter(ControllerConf config) {
-    return new MockControllerStarter(config);
+  protected ControllerStarter getControllerStarter() {
+    return new MockControllerStarter();
   }
 
   private class MockControllerStarter extends ControllerStarter {
     private static final int NUM_PERIODIC_TASKS = 8;
 
-    public MockControllerStarter(ControllerConf conf) {
-      super(conf);
+    public MockControllerStarter() {
+      super();
     }
 
     @Override
