+++ pinot-controller/src/test/java/org/apache/pinot/controller/helix/ControllerTest.java
@@ -75,6 +75,7 @@ import org.apache.pinot.spi.data.DimensionFieldSpec;
 import org.apache.pinot.spi.data.FieldSpec;
 import org.apache.pinot.spi.data.MetricFieldSpec;
 import org.apache.pinot.spi.data.Schema;
+import org.apache.pinot.spi.env.PinotConfiguration;
 import org.apache.pinot.spi.utils.CommonConstants;
 import org.apache.pinot.spi.utils.NetUtils;
 import org.apache.pinot.spi.utils.builder.TableNameBuilder;
@@ -150,11 +151,13 @@ public abstract class ControllerTest {
     return properties;
   }
 
-  protected void startController() {
+  protected void startController()
+      throws Exception {
     startController(getDefaultControllerConfiguration());
   }
 
-  protected void startController(Map<String, Object> properties) {
+  protected void startController(Map<String, Object> properties)
+      throws Exception {
     Preconditions.checkState(_controllerStarter == null);
 
     ControllerConf config = new ControllerConf(properties);
@@ -173,7 +176,8 @@ public abstract class ControllerTest {
     _controllerRequestURLBuilder = ControllerRequestURLBuilder.baseUrl(_controllerBaseApiUrl);
     _controllerDataDir = config.getDataDir();
 
-    _controllerStarter = getControllerStarter(config);
+    _controllerStarter = getControllerStarter();
+    _controllerStarter.init(new PinotConfiguration(properties));
     _controllerStarter.start();
     _helixResourceManager = _controllerStarter.getHelixResourceManager();
     _helixManager = _controllerStarter.getHelixControllerManager();
@@ -204,8 +208,8 @@ public abstract class ControllerTest {
     configAccessor.set(scope, CommonConstants.Helix.DEFAULT_HYPERLOGLOG_LOG2M_KEY, Integer.toString(12));
   }
 
-  protected ControllerStarter getControllerStarter(ControllerConf config) {
-    return new ControllerStarter(config);
+  protected ControllerStarter getControllerStarter() {
+    return new ControllerStarter();
   }
 
   protected int getControllerPort() {
