+++ pinot-integration-tests/src/test/java/org/apache/pinot/integration/tests/ExactlyOnceKafkaRealtimeClusterIntegrationTest.java
@@ -43,7 +43,8 @@ public class ExactlyOnceKafkaRealtimeClusterIntegrationTest extends RealtimeClus
   }
 
   @Override
-  public void startController() {
+  public void startController()
+      throws Exception {
     Map<String, Object> properties = getDefaultControllerConfiguration();
 
     properties.put(ControllerConf.ALLOW_HLC_TABLES, false);
