+++ pinot-plugins/pinot-minion-tasks/pinot-minion-builtin-tasks/src/test/java/org/apache/pinot/plugin/minion/tasks/segment_generation_and_push/SegmentGenerationAndPushTaskGeneratorTest.java
@@ -38,7 +38,8 @@ public class SegmentGenerationAndPushTaskGeneratorTest extends ControllerTest {
   SegmentGenerationAndPushTaskGenerator _generator;
 
   @BeforeClass
-  public void setup() {
+  public void setup()
+      throws Exception {
     int zkPort = 2171;
     startZk(zkPort);
     Map<String, Object> properties = getDefaultControllerConfiguration();
