+++ pinot-spi/src/main/java/org/apache/pinot/spi/services/ServiceStartable.java
@@ -27,6 +27,9 @@ import org.apache.pinot.spi.env.PinotConfiguration;
  *
  */
 public interface ServiceStartable {
+  void init(PinotConfiguration conf)
+      throws Exception;
+
   ServiceRole getServiceRole();
 
   String getInstanceId();
