+++ pinot-tools/src/main/java/org/apache/pinot/tools/perf/PerfBenchmarkDriver.java
@@ -188,15 +188,30 @@ public class PerfBenchmarkDriver {
     ZkStarter.startLocalZkServer(zkPort);
   }
 
-  private void startController() {
+  private void startController()
+      throws Exception {
     if (!_conf.shouldStartController()) {
       LOGGER.info("Skipping start controller step. Assumes controller is already started.");
       return;
     }
-    ControllerConf conf = getControllerConf();
+
     LOGGER.info("Starting controller at {}", _controllerAddress);
-    _controllerStarter = new ControllerStarter(conf);
+    _controllerStarter = new ControllerStarter();
+    _controllerStarter.init(new PinotConfiguration(getControllerProperties()));
     _controllerStarter.start();
+  }
+
+  private Map<String, Object> getControllerProperties() {
+    Map<String, Object> properties = new HashMap<>();
+    properties.put(CommonConstants.Helix.CONFIG_OF_CLUSTER_NAME, _clusterName);
+    properties.put(CommonConstants.Helix.CONFIG_OF_ZOOKEEPR_SERVER, _zkAddress);
+    properties.put(ControllerConf.CONTROLLER_HOST, _controllerHost);
+    properties.put(ControllerConf.CONTROLLER_PORT, String.valueOf(_controllerPort));
+    properties.put(ControllerConf.DATA_DIR, _controllerDataDir);
+    properties.put(ControllerConf.CLUSTER_TENANT_ISOLATION_ENABLE, false);
+    properties.put(ControllerConf.CONTROLLER_VIP_HOST, "localhost");
+    properties.put(ControllerConf.CONTROLLER_VIP_PROTOCOL, CommonConstants.HTTP_PROTOCOL);
+    return properties;
   }
 
   private ControllerConf getControllerConf() {
@@ -228,7 +243,9 @@ public class PerfBenchmarkDriver {
 
     LOGGER.info("Starting broker instance: {}", brokerInstanceName);
 
-    new HelixBrokerStarter(new PinotConfiguration(properties)).start();
+    HelixBrokerStarter helixBrokerStarter = new HelixBrokerStarter();
+    helixBrokerStarter.init(new PinotConfiguration(properties));
+    helixBrokerStarter.start();
   }
 
   private void startServer()
@@ -251,7 +268,8 @@ public class PerfBenchmarkDriver {
 
     LOGGER.info("Starting server instance: {}", _serverInstanceName);
 
-    HelixServerStarter helixServerStarter = new HelixServerStarter(new PinotConfiguration(properties));
+    HelixServerStarter helixServerStarter = new HelixServerStarter();
+    helixServerStarter.init(new PinotConfiguration(properties));
     helixServerStarter.start();
   }
